Creating TestPaymentType file in TestBangazonApi. Adding test to get all payment types and verified that test ran successfully.
