--- TestBangazonAPI/TestCustomers.cs
@@ -1,40 +1,0 @@
-using System;
-using System.Net;
-using Newtonsoft.Json;
-using Xunit;
-using BangazonAPI.Models;
-using System.Threading.Tasks;
-using System.Collections.Generic;
-
-namespace TestBangazonAPI
-{
-    public class TestCustomers
-    {
-        [Fact]
-        public async Task Test_Get_All_Customers()
-        {
-            using (var client = new APIClientProvider().Client)
-            {
-                /*
-                    ARRANGE
-                */
-
-
-                /*
-                    ACT
-                */
-                var response = await client.GetAsync("/api/customers");
-
-
-                string responseBody = await response.Content.ReadAsStringAsync();
-                var customers = JsonConvert.DeserializeObject<List<Customer>>(responseBody);
-
-                /*
-                    ASSERT
-                */
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-                Assert.True(customers.Count > 0);
-            }
-        }
-    }
-}
